requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-tinymce==2.8.0
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-decouple==3.1
pytz==2018.9
whitenoise==4.1.2
